+++ experimental/dds/tree2/src/feature-libraries/editable-tree-2/proxies/index.ts
@@ -16,9 +16,5 @@ export {
 	ProxyRoot,
 	SharedTreeNode,
 } from "./types";
-export {
-	StructFactorySchemaBuilder,
-	SharedTreeObjectFactory,
-	TreeNodeSchemaWithObjectFactory,
-} from "./schemaBuilder";
+export { SharedTreeObjectFactory, FactoryTreeSchema, addFactory } from "./objectFactory";
 export { node } from "./node";
